Добавил переменные в index.html
