+++ .gitignore
@@ -160,4 +160,5 @@ cython_debug/
 #.idea/
 
 # Custom
-.config.py
+config.py
+*.db
