+++ create_db.py
@@ -1,4 +1,4 @@
-from webapp import db, create_app
+from app import db, create_app
 
 
 with create_app().app_context():
