--- webapp/__init__.py
@@ -1,16 +1,0 @@
-from flask import Flask, render_template
-
-from webapp.model import db
-
-
-def create_app():
-    app = Flask(__name__)
-    app.config.from_pyfile('config.py')
-    db.init_app(app)
-
-    @app.route('/')
-    def index():
-        title = 'Свежее'
-        return render_template('index.html')
-    
-    return app
