--- webapp/templates/base.html
@@ -1,112 +1,0 @@
-<!doctype html>
-<html lang="ru">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Cooking</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-  <link href="../static/css/style.css" rel="stylesheet">
-</head>
-<body>
-
-  <header class="p-3 mb-3 border-bottom">
-    <div class="container">
-      <div class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start">
-        <div class="col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
-          <a href="/" class="d-flex align-items-center link-body-emphasis text-decoration-none">
-            <b>Лого</b>
-          </a>
-        </div>  
-
-        <div class="dropdown text-end">
-          <a href="#" class="d-block link-body-emphasis text-decoration-none dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">
-            <img src="../static/img/no_photo.png" alt="photo" class="rounded-circle" width="64" height="64">
-          </a>
-          <ul class="dropdown-menu text-small">
-            <li><a class="dropdown-item" href="#">Мои рецепты</a></li>
-            <li><a class="dropdown-item" href="#">Добавить рецепт</a></li>
-            <li><a class="dropdown-item" href="#">Избранное</a></li>
-            <li><hr class="dropdown-divider"></li>
-            <li><a class="dropdown-item" href="#">Выход</a></li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </header>
-
-  <div class="container">
-    <div class="row">
-      <div class="col-md-2 col-sm-12 mb-3">
-        <div class="sidebar border rounded-3 p-2">
-          <ul class="nav nav-pills flex-column mb-auto text-center">
-            <li class="nav-item"><a href="#" class="nav-link active" aria-current="page">Свежее</a></li>
-            <li><a href="#" class="nav-link link-body-emphasis">Топ</a></li>
-            <hr>
-            <li><a href="#" class="nav-link link-body-emphasis">Поиск</a></li>
-          </ul>
-        </div>
-      </div>
-
-      <div class="col-md-10 col-sm-12">
-        <div class="posts-header border rounded-3 p-2 text-center mb-3">
-          <h5>Свежее</h5>
-        </div>
-        <div class="rounded-1 p-1">
-          <div class="card mb-3">
-            <div class="card-header">
-            <h5>Название рецепта</h5>
-            <span>Автор: Вася Пупкин</span>
-            </div>
-            <div class="card-body">
-              <span><img class="img-thumbnail rounded float-start me-4" src="../static/img/posts_img.png" alt="posts_img" width="180px" height="100px"></span>
-              <p class="card-text">Краткое описание</p>
-            </div>
-            <div class="card-footer text-body-secondary">
-              <span class="">Дата и время публикации</span>
-              <a href="#" class="btn btn-success">Редактировать</a>
-              <a href="#" class="btn btn-danger">Удалить</a>
-              <a href="#" class="btn btn-primary">Читать</a>
-            </div>
-          </div>
-          <div class="card mb-3">
-            <div class="card-header">
-            <h5>Название рецепта</h5>
-            <span>Автор: Вася Пупкин</span>
-            </div>
-            <div class="card-body">
-              <span><img class="img-thumbnail rounded float-start me-4" src="../static/img/posts_img.png" alt="posts_img" width="180px" height="100px"></span>
-              <p class="card-text">Краткое описание</p>
-            </div>
-            <div class="card-footer text-body-secondary">
-              <span class="">Дата и время публикации</span>
-              <a href="#" class="btn btn-success">Редактировать</a>
-              <a href="#" class="btn btn-danger">Удалить</a>
-              <a href="#" class="btn btn-primary">Читать</a>
-            </div>
-          </div>
-          <div class="card mb-3">
-            <div class="card-header">
-            <h5>Название рецепта</h5>
-            <span>Автор: Вася Пупкин</span>
-            </div>
-            <div class="card-body">
-              <span><img class="img-thumbnail rounded float-start me-4" src="../static/img/posts_img.png" alt="posts_img" width="180px" height="100px"></span>
-              <p class="card-text">Краткое описание</p>
-            </div>
-            <div class="card-footer text-body-secondary">
-              <span class="">Дата и время публикации</span>
-              <a href="#" class="btn btn-success">Редактировать</a>
-              <a href="#" class="btn btn-danger">Удалить</a>
-              <a href="#" class="btn btn-primary">Читать</a>
-            </div>
-          </div>
-          </div>
-        </div>
-      </div>
-
-    </div>
-  </div>
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"></script>
-</body>
-</html>
